🚚move : assets을 public으로 이동
